requirements.txt
# Phase 1: Trend Scraping
apify-client==2.5.0        # Apify SDK for TikTok scraping
python-dotenv==1.2.2       # Load .env files
rich==14.3.3               # Pretty terminal output

# Phase 2: Script Generation
anthropic==0.84.0          # Claude API SDK

# Testing
pytest==9.0.2

# Phase 3: Video Rendering
ffmpeg-python==0.2.0       # FFmpeg Python bindings

# Phase 4: TikTok Publishing
requests==2.32.5           # HTTP client for TikTok API

# Dashboard
fastapi[standard]==0.115.12  # Web dashboard (includes uvicorn, jinja2)

# Product Video Pipeline
google-genai==1.20.0         # Gemini API for image generation (Nano Banana Pro)
playwright==1.52.0           # Browser automation for Kalodata scraping
# Muapi video generation uses requests (already installed above) — no SDK needed

# Video Analysis Pipeline
Pillow==11.2.1               # Image resizing for frame encoding
